+++ jaden-backend/data/users.json
@@ -1,0 +1,5 @@
+{
+    "username": "Michael",
+    "email": "michael@me.com",
+    "password": "password"
+}
