+++ jaden-backend/db.js
@@ -1,0 +1,8 @@
+const { DataStore } = require('notarealdb')
+
+const db = new DataStore('./data')
+
+module.exports = {
+    users: db.collection('users'),
+    tours: db.collection('tours')
+}
