+++ jaden-backend/package.json
@@ -10,11 +10,16 @@
   "license": "ISC",
   "dependencies": {
     "apollo-server": "^2.19.1",
+    "apollo-server-express": "^2.19.1",
+    "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-graphql": "^0.12.0",
+    "express-jwt": "^6.0.0",
     "graphql": "^15.4.0",
-    "mongoose": "^5.11.4"
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.11.4",
+    "notarealdb": "^0.2.2"
   },
   "devDependencies": {
     "dotenv": "^8.2.0",
